requirements.txt
asgiref==3.7.2
Django==4.2.4
gunicorn==21.2.0
django-environ==0.10.0
psycopg2-binary==2.9.6
Pillow==10.0.0
rpi.gpio==0.7.1
gpiozero==1.6.2
Redis==4.6.0
celery==5.3.1
channels[daphne]==4.0.0
channels_redis==4.1.0